Move CSourceLayer, CSinkLayer and CBaseInPlaceLayer to separate files
