+++ NeoML/include/NeoML/Dnn/Layers/ActivationLayers.h
@@ -17,6 +17,7 @@ limitations under the License.
 
 #include <NeoML/NeoMLDefs.h>
 #include <NeoML/Dnn/Dnn.h>
+#include <NeoML/Dnn/Layers/BaseInPlaceLayer.h>
 #include <NeoMathEngine/NeoMathEngine.h>
 
 namespace NeoML {
