+++ NeoML/include/NeoML/Dnn/Layers/SinkLayer.h
@@ -1,4 +1,4 @@
-﻿/* Copyright © 2017-2020 ABBYY Production LLC
+/* Copyright © 2017-2020 ABBYY Production LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,38 +17,28 @@ limitations under the License.
 
 #include <NeoML/NeoMLDefs.h>
 #include <NeoML/Dnn/Dnn.h>
-#include <NeoML/Dnn/Layers/BaseInPlaceLayer.h>
 
 namespace NeoML {
 
-// Implements the layer that calculates the softmax function. 
-// Note that if you are using cross-entropy loss function you may set softmax to be calculated there and will not need this layer.
-class NEOML_API CSoftmaxLayer : public CBaseInPlaceLayer {
-	NEOML_DNN_LAYER( CSoftmaxLayer )
+// The layer that is used to pass a data blob out of the network
+class NEOML_API CSinkLayer : public CBaseLayer {
+	NEOML_DNN_LAYER( CSinkLayer )
 public:
-	explicit CSoftmaxLayer( IMathEngine& mathEngine );
-
-	// The area (dimensions) over which the values are normalized
-	enum TNormalizationArea {
-		NA_ObjectSize = 0,
-		NA_BatchLength,
-		NA_ListSize,
-		NA_Channel,
-
-		NA_Count
-	};
-
-	void SetNormalizationArea( TNormalizationArea newArea ) { area = newArea; }
-	TNormalizationArea GetNormalizationArea() const { return area; }
+	explicit CSinkLayer( IMathEngine& mathEngine ) : CBaseLayer( mathEngine, "CCnnSinkLayer", false ) {}
 
 	void Serialize( CArchive& archive ) override;
 
+	// Gets the reference to the output blob
+	// It is valid only after the RunOnce method called
+	// After each call to RunOnce this blob contains the results
+	const CPtr<CDnnBlob>& GetBlob() const;
+
 protected:
+	CPtr<CDnnBlob> blob;
+
+	void Reshape() override;
 	void RunOnce() override;
 	void BackwardOnce() override;
-
-private:
-	TNormalizationArea area; // the normalization area
 };
 
 } // namespace NeoML
